test(ui): share focused sidebar drag data fixture
